terminate this repo, move on to ver-01-2016
summary:
this version is a minor update from ver-07-2015. Included features are
slanted matrix crack, and predelam elset.

move onto ver-01-2016, targeted changes are:
- all fnm elements form one part in the model
- all bcds/constraints on the fnm elements are applied on the nsests of
the fnm part, NOT the nsests of the final assembly.
- fl. nodes affected by the bcds will be taken care of in the element
subroutine.
These changes would allow the combined use of both fnm elements and the
default abaqus elements. fnm part can be tied to the default abaqus
parts. This allows more capabilities to be unleashed, more problems to
be solved.